+++ geo/CHANGES.md
@@ -15,6 +15,8 @@
 - Renamed features `use-proj`, `use-serde` to simply `proj` and `serde` (removing the `use-` prefix)
   and deprecated the old spelling.
   - <https://github.com/georust/geo/pull/1447>
+- Update `ConcaveHull` algorithm with implementation of [mapbox/concaveman](https://github.com/mapbox/concaveman).
+- BREAKING: The `concave_hull` method now requires a `length_threshold` parameter.
 
 ## 0.31.0 - 2025-09-01